+++ src/Mount.cpp
@@ -8,7 +8,8 @@
 PUSH_NO_WARNINGS
 #include <AccelStepper.h>
 #if (RA_DRIVER_TYPE == DRIVER_TYPE_TMC2209_UART) || (DEC_DRIVER_TYPE == DRIVER_TYPE_TMC2209_UART)                                          \
-    || (AZ_DRIVER_TYPE == DRIVER_TYPE_TMC2209_UART) || (ALT_DRIVER_TYPE == DRIVER_TYPE_TMC2209_UART)
+    || (AZ_DRIVER_TYPE == DRIVER_TYPE_TMC2209_UART) || (ALT_DRIVER_TYPE == DRIVER_TYPE_TMC2209_UART)                                       \
+    || (FOCUS_DRIVER_TYPE == DRIVER_TYPE_TMC2209_UART)
     #include <TMCStepper.h>  // If you get an error here, download the TMCstepper library from "Tools > Manage Libraries"
 #endif
 POP_NO_WARNINGS
@@ -319,7 +320,11 @@ void Mount::configureDECStepper(byte pin1, byte pin2, int maxSpeed, int maxAccel
     #if AZ_DRIVER_TYPE == DRIVER_TYPE_ULN2003
 void Mount::configureAZStepper(byte pin1, byte pin2, byte pin3, byte pin4, int maxSpeed, int maxAcceleration)
 {
-    _stepperAZ = new AccelStepper((AZ_MICROSTEPPING == 1) ? AccelStepper::FULL4WIRE : AccelStepper::HALF4WIRE, pin1, pin2, pin3, pin4);
+        #if AZ_INVERT_DIR == 1
+    _stepperAZ = new AccelStepper((AZ_MICROSTEPPING == 1) ? AccelStepper::FULL4WIRE : AccelStepper::HALF4WIRE, pin4, pin3, pin2, pin1);
+        #else
+    _stepperAZ  = new AccelStepper((AZ_MICROSTEPPING == 1) ? AccelStepper::FULL4WIRE : AccelStepper::HALF4WIRE, pin1, pin2, pin3, pin4);
+        #endif
     _stepperAZ->setSpeed(0);
     _stepperAZ->setMaxSpeed(maxSpeed);
     _stepperAZ->setAcceleration(maxAcceleration);
@@ -334,6 +339,9 @@ void Mount::configureAZStepper(byte pin1, byte pin2, int maxSpeed, int maxAccele
     _stepperAZ->setAcceleration(maxAcceleration);
     _maxAZSpeed        = maxSpeed;
     _maxAZAcceleration = maxAcceleration;
+        #if AZ_INVERT_DIR == 1
+    _stepperAZ->setPinsInverted(true, false, false);
+        #endif
 }
     #endif
 #endif
@@ -341,7 +349,11 @@ void Mount::configureAZStepper(byte pin1, byte pin2, int maxSpeed, int maxAccele
     #if ALT_DRIVER_TYPE == DRIVER_TYPE_ULN2003
 void Mount::configureALTStepper(byte pin1, byte pin2, byte pin3, byte pin4, int maxSpeed, int maxAcceleration)
 {
+        #if ALT_INVERT_DIR == 1
+    _stepperALT = new AccelStepper((ALT_MICROSTEPPING == 1) ? AccelStepper::FULL4WIRE : AccelStepper::HALF4WIRE, pin4, pin3, pin2, pin1);
+        #else
     _stepperALT = new AccelStepper((ALT_MICROSTEPPING == 1) ? AccelStepper::FULL4WIRE : AccelStepper::HALF4WIRE, pin1, pin2, pin3, pin4);
+        #endif
     _stepperALT->setSpeed(0);
     _stepperALT->setMaxSpeed(maxSpeed);
     _stepperALT->setAcceleration(maxAcceleration);
@@ -356,6 +368,9 @@ void Mount::configureALTStepper(byte pin1, byte pin2, int maxSpeed, int maxAccel
     _stepperALT->setAcceleration(maxAcceleration);
     _maxALTSpeed        = maxSpeed;
     _maxALTAcceleration = maxAcceleration;
+        #if ALT_INVERT_DIR == 1
+    _stepperALT->setPinsInverted(true, false, false);
+        #endif
 }
     #endif
 #endif
@@ -395,7 +410,8 @@ void Mount::configureFocusStepper(byte pin1, byte pin2, int maxSpeed, int maxAcc
 #endif
 
 #if RA_DRIVER_TYPE == DRIVER_TYPE_TMC2209_UART || DEC_DRIVER_TYPE == DRIVER_TYPE_TMC2209_UART                                              \
-    || AZ_DRIVER_TYPE == DRIVER_TYPE_TMC2209_UART || ALT_DRIVER_TYPE == DRIVER_TYPE_TMC2209_UART
+    || AZ_DRIVER_TYPE == DRIVER_TYPE_TMC2209_UART || ALT_DRIVER_TYPE == DRIVER_TYPE_TMC2209_UART                                           \
+    || FOCUS_DRIVER_TYPE == DRIVER_TYPE_TMC2209_UART
     #if UART_CONNECTION_TEST_TXRX == 1
 bool Mount::connectToDriver(TMC2209Stepper *driver, const char *driverKind)
 {
@@ -772,8 +788,8 @@ void Mount::configureFocusDriver(Stream *serial, float rsense, byte driveraddres
         #if USE_VREF == 0  //By default, Vref is ignored when using UART to specify rms current.
     _driverFocus->I_scale_analog(false);
         #endif
-    LOGV2(DEBUG_STEPPERS, F("Mount: Requested Focus motor rms_current: %d mA"), rmscurrent);
-    _driverFocus->rms_current(rmscurrent, 1.0f);  //holdMultiplier = 1 to set ihold = irun
+    LOGV2(DEBUG_STEPPERS | DEBUG_FOCUS, F("Mount: Requested Focus motor rms_current: %d mA"), rmscurrent);
+    _driverFocus->rms_current(rmscurrent, FOCUSER_MOTOR_HOLD_SETTING / 100.f);  //holdMultiplier = 1 to set ihold = irun
     _driverFocus->toff(1);
     _driverFocus->en_spreadCycle(FOCUS_UART_STEALTH_MODE == 0);
     _driverFocus->blank_time(24);
@@ -784,10 +800,15 @@ void Mount::configureFocusDriver(Stream *serial, float rsense, byte driveraddres
         #if UART_CONNECTION_TEST_TXRX == 1
     if (UART_Rx_connected)
     {
-        LOGV2(DEBUG_STEPPERS, F("Mount: Actual Focus motor rms_current: %d mA"), _driverFocus->rms_current());
-        LOGV2(DEBUG_STEPPERS, F("Mount: Actual Focus CS value: %d"), _driverFocus->cs_actual());
-        LOGV2(DEBUG_STEPPERS, F("Mount: Actual Focus vsense: %d"), _driverFocus->vsense());
+        LOGV2(DEBUG_STEPPERS | DEBUG_FOCUS, F("Mount: Actual Focus motor rms_current: %d mA"), _driverFocus->rms_current());
+        LOGV2(DEBUG_STEPPERS | DEBUG_FOCUS, F("Mount: Actual Focus CS value: %d"), _driverFocus->cs_actual());
+        LOGV2(DEBUG_STEPPERS | DEBUG_FOCUS, F("Mount: Actual Focus vsense: %d"), _driverFocus->vsense());
     }
+        #endif
+
+        #if FOCUSER_ALWAYS_ON == 1
+    LOGV1(DEBUG_FOCUS, F("Mount::configureFocusDriver: Always on -> TMC2209U enabling driver pin."));
+    digitalWrite(FOCUS_EN_PIN, LOW);  // Logic LOW to enable driver
         #endif
 }
 
@@ -813,8 +834,8 @@ void Mount::configureFocusDriver(
         #if USE_VREF == 0  //By default, Vref is ignored when using UART to specify rms current.
     _driverFocus->I_scale_analog(false);
         #endif
-    LOGV2(DEBUG_STEPPERS, F("Mount: Requested Focus motor rms_current: %d mA"), rmscurrent);
-    _driverFocus->rms_current(rmscurrent, 1.0f);  //holdMultiplier = 1 to set ihold = irun
+    LOGV2(DEBUG_STEPPERS | DEBUG_FOCUS, F("Mount: Requested Focus motor rms_current: %d mA"), rmscurrent);
+    _driverFocus->rms_current(rmscurrent, FOCUSER_MOTOR_HOLD_SETTING / 100.f);  //holdMultiplier = 1 to set ihold = irun
     _driverFocus->toff(1);
     _driverFocus->en_spreadCycle(FOCUS_UART_STEALTH_MODE == 0);
     _driverFocus->blank_time(24);
@@ -825,10 +846,14 @@ void Mount::configureFocusDriver(
         #if UART_CONNECTION_TEST_TXRX == 1
     if (UART_Rx_connected)
     {
-        LOGV2(DEBUG_STEPPERS, F("Mount: Actual Focus motor rms_current: %d mA"), _driverFocus->rms_current());
-        LOGV2(DEBUG_STEPPERS, F("Mount: Actual Focus CS value: %d"), _driverFocus->cs_actual());
-        LOGV2(DEBUG_STEPPERS, F("Mount: Actual Focus vsense: %d"), _driverFocus->vsense());
+        LOGV2(DEBUG_STEPPERS | DEBUG_FOCUS, F("Mount: Actual Focus motor rms_current: %d mA"), _driverFocus->rms_current());
+        LOGV2(DEBUG_STEPPERS | DEBUG_FOCUS, F("Mount: Actual Focus CS value: %d"), _driverFocus->cs_actual());
+        LOGV2(DEBUG_STEPPERS | DEBUG_FOCUS, F("Mount: Actual Focus vsense: %d"), _driverFocus->vsense());
     }
+        #endif
+        #if FOCUSER_ALWAYS_ON == 1
+    LOGV1(DEBUG_FOCUS, F("Mount::configureFocusDriver: Always on -> TMC2209U enabling driver pin."));
+    digitalWrite(FOCUS_EN_PIN, LOW);  // Logic LOW to enable driver
         #endif
 }
     #endif
@@ -1708,7 +1733,7 @@ void Mount::setSpeed(StepperAxis which, float speedDegsPerSec)
 #if (FOCUS_STEPPER_TYPE != STEPPER_TYPE_NONE)
     else if (which == FOCUS_STEPS)
     {
-        LOGV2(DEBUG_MOUNT, F("Mount: Focuser setSpeed %f"), speedDegsPerSec);
+        LOGV2(DEBUG_MOUNT | DEBUG_FOCUS, F("Mount:setSpeed() Focuser setSpeed %f"), speedDegsPerSec);
 
     #if FOCUS_DRIVER_TYPE == DRIVER_TYPE_ULN2003
         float curFocusSpeed = _stepperFocus->speed();
@@ -1736,7 +1761,7 @@ void Mount::setSpeed(StepperAxis which, float speedDegsPerSec)
         || FOCUS_DRIVER_TYPE == DRIVER_TYPE_TMC2209_UART
         if (speedDegsPerSec != 0)
         {
-            LOGV2(DEBUG_STEPPERS, F("STEP-setSpeed: Enabling motor and setting speed. Continuous"), speedDegsPerSec);
+            LOGV2(DEBUG_STEPPERS | DEBUG_FOCUS, F("Mount:setSpeed(): Enabling motor, setting speed to %f. Continuous"), speedDegsPerSec);
             enableFocusMotor();
             _stepperFocus->setMaxSpeed(speedDegsPerSec);
             _stepperFocus->moveTo(sign(speedDegsPerSec) * 300000);
@@ -1744,7 +1769,7 @@ void Mount::setSpeed(StepperAxis which, float speedDegsPerSec)
         }
         else
         {
-            LOGV2(DEBUG_STEPPERS, F("STEP-setSpeed: Stopping motor."), speedDegsPerSec);
+            LOGV1(DEBUG_STEPPERS | DEBUG_FOCUS, F("Mount:setSpeed(): Stopping motor."));
             _stepperFocus->stop();
         }
 
@@ -1903,7 +1928,7 @@ void Mount::enableAzAltMotors()
         #if AZ_DRIVER_TYPE == DRIVER_TYPE_ULN2003
     _stepperAZ->enableOutputs();
         #else
-    digitalWrite(AZ_EN_PIN, LOW);      // Logic LOW to enable driver
+    digitalWrite(AZ_EN_PIN, LOW);   // Logic LOW to enable driver
         #endif
     #endif
 
@@ -1911,7 +1936,7 @@ void Mount::enableAzAltMotors()
         #if ALT_DRIVER_TYPE == DRIVER_TYPE_ULN2003
     _stepperALT->enableOutputs();
         #else
-    digitalWrite(ALT_EN_PIN, LOW);     // Logic LOW to enable driver
+    digitalWrite(ALT_EN_PIN, LOW);  // Logic LOW to enable driver
         #endif
     #endif
 }
@@ -1949,12 +1974,17 @@ void Mount::focusSetSpeedByRate(int rate)
     _focusRate          = clamp(rate, 1, 4);
     float speedFactor[] = {0, 0.05, 0.2, 0.5, 1.0};
     _maxFocusRateSpeed  = speedFactor[_focusRate] * _maxFocusSpeed;
-    LOGV3(DEBUG_MOUNT, F("Mount::focusSetSpeedByRate: rate is %d -> %f"), _focusRate, _maxFocusRateSpeed);
+    LOGV5(DEBUG_FOCUS,
+          F("Mount::focusSetSpeedByRate: rate is %d, factor %f, maxspeed %f -> %f"),
+          _focusRate,
+          speedFactor[_focusRate],
+          _maxFocusSpeed,
+          _maxFocusRateSpeed);
     _stepperFocus->setMaxSpeed(_maxFocusRateSpeed);
 
     if (_stepperFocus->isRunning())
     {
-        LOGV1(DEBUG_MOUNT, F("Mount::focusSetSpeedByRate: stepper is already running so adjust speed"));
+        LOGV1(DEBUG_FOCUS, F("Mount::focusSetSpeedByRate: stepper is already running so should adjust speed?"));
         //_stepperFocus->setSpeed(speedFactor[_focusRate ] * _maxFocusSpeed);
     }
 }
@@ -1967,6 +1997,7 @@ void Mount::focusSetSpeedByRate(int rate)
 void Mount::focusContinuousMove(FocuserDirection direction)
 {
     // maxSpeed is set to what the rate dictates
+    LOGV3(DEBUG_FOCUS, F("Mount::focusContinuousMove: direction is %d, maxspeed %f"), direction, _maxFocusRateSpeed);
     setSpeed(FOCUS_STEPS, static_cast<int>(direction) * _maxFocusRateSpeed);
 }
 
@@ -1978,7 +2009,7 @@ void Mount::focusContinuousMove(FocuserDirection direction)
 void Mount::focusMoveBy(long steps)
 {
     long targetPosition = _stepperFocus->currentPosition() + steps;
-    LOGV3(DEBUG_MOUNT, F("Mount::focusMoveBy: move by %l steps to %l. Target Mode."), steps, targetPosition);
+    LOGV3(DEBUG_FOCUS, F("Mount::focusMoveBy: move by %l steps to %l. Target Mode."), steps, targetPosition);
     enableFocusMotor();
     _stepperFocus->moveTo(targetPosition);
     _focuserMode = FOCUS_TO_TARGET;
@@ -2011,16 +2042,25 @@ void Mount::focusSetStepperPosition(long steps)
 /////////////////////////////////
 void Mount::disableFocusMotor()
 {
-    LOGV1(DEBUG_MOUNT, F("Mount::disableFocusMotor: stopping motor and waiting."));
+    LOGV1(DEBUG_FOCUS, F("Mount::disableFocusMotor: stopping motor and waiting."));
     _stepperFocus->stop();
     waitUntilStopped(FOCUSING);
 
-    LOGV1(DEBUG_MOUNT, F("Mount::disableFocusMotor: disabling motor"));
     #if (FOCUS_STEPPER_TYPE != STEPPER_TYPE_NONE)
         #if FOCUS_DRIVER_TYPE == DRIVER_TYPE_ULN2003
+    LOGV1(DEBUG_FOCUS, F("Mount::disableFocusMotor: ULN2003 disabling outputs"));
     _stepperFocus->disableOutputs();
         #else
+            #if FOCUSER_ALWAYS_ON == 0
+                #if (FOCUS_DRIVER_TYPE == DRIVER_TYPE_TMC2209_UART)
+
+    LOGV1(DEBUG_FOCUS, F("Mount::disableFocusMotor: TMC2209U disabling enable pin"));
     digitalWrite(FOCUS_EN_PIN, HIGH);  // Logic HIGH to disable driver
+                #else
+    LOGV1(DEBUG_FOCUS, F("Mount::disableFocusMotor: non-TMC2209U disabling enable pin"));
+    digitalWrite(FOCUS_EN_PIN, HIGH);  // Logic HIGH to disable driver
+                #endif
+            #endif
         #endif
     #endif
 }
@@ -2032,10 +2072,11 @@ void Mount::disableFocusMotor()
 /////////////////////////////////
 void Mount::enableFocusMotor()
 {
-    LOGV1(DEBUG_MOUNT, F("Mount::enableFocusMotor: enabling."));
     #if FOCUS_DRIVER_TYPE == DRIVER_TYPE_ULN2003
+    LOGV1(DEBUG_FOCUS, F("Mount::enableFocusMotor: ULN2003 enabling outputs."));
     _stepperFocus->enableOutputs();
     #else
+    LOGV1(DEBUG_FOCUS, F("Mount::enableFocusMotor: enabling driver pin."));
     digitalWrite(FOCUS_EN_PIN, LOW);  // Logic LOW to enable driver
     #endif
 }
@@ -2047,7 +2088,7 @@ void Mount::enableFocusMotor()
 /////////////////////////////////
 void Mount::focusStop()
 {
-    LOGV1(DEBUG_MOUNT, F("Mount::focusStop: stopping motor."));
+    LOGV1(DEBUG_FOCUS, F("Mount::focusStop: stopping motor."));
     _stepperFocus->stop();
 }
 
@@ -2801,19 +2842,18 @@ void Mount::loop()
 
     if (_stepperFocus->isRunning())
     {
-        LOGV2(DEBUG_MOUNT, F("Mount: Focuser running at speed %f"), _stepperFocus->speed());
+        LOGV2(DEBUG_FOCUS, F("Mount:loop(): Focuser running at speed %f"), _stepperFocus->speed());
         _focuserWasRunning = true;
     }
     else if (_focuserWasRunning)
     {
-        LOGV1(DEBUG_MOUNT, F("Mount: Focuser is stopped, but was running "));
+        LOGV1(DEBUG_FOCUS, F("Mount:loop(): Focuser is stopped, but was running "));
         // If focuser was running last time through the loop, but not this time, it has
         // either been stopped, or reached the target.
         _focuserMode       = FOCUS_IDLE;
         _focuserWasRunning = false;
-    #if FOCUSER_ALWAYS_ON == 0
+        LOGV1(DEBUG_FOCUS, F("Mount:loop(): Focuser is stopped, but was running, disabling"));
         disableFocusMotor();
-    #endif
     }
 
 #endif
@@ -3114,7 +3154,7 @@ void Mount::setHome(bool clearZeroPos)
     //LOGV2(DEBUG_MOUNT_VERBOSE,F("Mount::setHomePre: currentRA is %s"), currentRA().ToString());
     //LOGV2(DEBUG_MOUNT_VERBOSE,F("Mount::setHomePre: targetRA is %s"), targetRA().ToString());
     //LOGV2(DEBUG_MOUNT_VERBOSE,F("Mount::setHomePre: zeroPos is %s"), _zeroPosRA.ToString());
-    _zeroPosRA = clearZeroPos ? DayTime(POLARIS_RA_HOUR, POLARIS_RA_MINUTE, POLARIS_RA_SECOND) : currentRA();
+    _zeroPosRA = clearZeroPos ? DayTime(POLARIS_RA_HOUR, POLARIS_RA_MINUTE, POLARIS_RA_SECOND) : calculateLst();
 
     _stepperRA->setCurrentPosition(0);
     _stepperDEC->setCurrentPosition(0);